OneToMany; check user

+++ src/tasks/tasks.service.ts
@@ -4,6 +4,8 @@ import { CreateTaskDto } from './dto/create-task.dto';
 import { TaskRepo } from './repos/task.repo';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './entities/task.entity';
+import { User } from 'src/auth/entities/user.entity';
+import { FilterTasksDto } from './dto/filter-tasks.dto';
 
 @Injectable()
 export class TasksService {
@@ -12,31 +14,37 @@ export class TasksService {
     private taskRepo: TaskRepo,
   ) {}
 
-  getTasks(filterTasksDto): Promise<Task[]> {
-    return this.taskRepo.getTasks(filterTasksDto);
+  getTasks(filterTasksDto: FilterTasksDto, user: User): Promise<Task[]> {
+    return this.taskRepo.getTasks(filterTasksDto, user);
   }
 
-  async getTaksById(id: number): Promise<Task> {
-    const task = await this.taskRepo.findOne(id);
+  async getTaksById(id: number, user: User): Promise<Task> {
+    const task = await this.taskRepo.findOne({
+      where: { id, userId: user.id },
+    });
     if (!task) {
       throw new NotFoundException(`Task with ${id} not found`);
     }
     return task;
   }
 
-  createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepo.createTask(createTaskDto);
+  createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepo.createTask(createTaskDto, user);
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepo.delete(id);
+  async deleteTask(id: number, user: User): Promise<void> {
+    const result = await this.taskRepo.delete({ id, userId: user.id });
     if (result.affected === 0) {
       throw new NotFoundException();
     }
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaksById(id);
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaksById(id, user);
     task.status = status;
     task.save();
     return task;
